Give shared exam grading policy a non-presentation owner (#572)
Move the seven shared exam-grading symbols (grade_item,
grade_objective_item, grade_short_answer_item, build_grading_result,
build_tracking_update, build_exam_summary, TERMINAL_GRADING_STATUSES) to
backend/app/exam_grading.py. Routes, workers, and direct tests now import
from the new non-presentation owner. The exhausted
app/presentation/exam_grading.py module is deleted and the two moved
symbols are removed from exam_helpers.py. Implementations, signatures,
grading parity, and lifecycle behavior are unchanged.
